pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmesh-llive"
version = "0.6.0"
description = "Self-evolving modular memory LLM framework — biological memory × formal verification × industrial IoT mesh × TUI HITL"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE", "LICENSE-COMMERCIAL", "NOTICE"]
requires-python = ">=3.11,<3.12"
authors = [{ name = "Kazufumi Furuse", email = "kazufumi@furuse.work" }]
keywords = [
  "llm",
  "self-evolution",
  "modular-memory",
  "continual-learning",
  "neural-architecture-search",
  "industrial-iot",
  "tui",
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Operating System :: OS Independent",
]
dependencies = [
  "jsonschema>=4.21",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "numpy>=1.26",
  "duckdb>=0.10",
  "structlog>=24.1",
  "typer>=0.12",
  "kuzu>=0.4",
  "apscheduler>=3.10",
  "safetensors>=0.4",
  "cryptography>=42.0",
]

[project.optional-dependencies]
torch = [
  "torch>=2.2",
  "transformers>=4.40",
  "accelerate>=0.30",
  "sentence-transformers>=2.7",
  "faiss-cpu>=1.8",
  "peft>=0.10",
  "hdbscan>=0.8",
]
ingest = [
  "pypdf>=4.0",
  "arxiv>=2.1",
  "readability-lxml>=0.8",
  "requests>=2.31",
]
llm = [
  "anthropic>=0.30",
]
verify = [
  "z3-solver>=4.13",
]
rust = [
  "maturin>=1.5",
]
mcp = [
  "mcp>=1.0",
]
vlm = [
  "pillow>=10.0",
]
coding = [
]
dev = [
  "pytest>=8",
  "pytest-cov>=4.1",
  "hypothesis>=6.100",
  "ruff>=0.4",
  "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/furuse-kazufumi/llive"
Documentation = "https://github.com/furuse-kazufumi/llive/tree/main/docs"
Issues = "https://github.com/furuse-kazufumi/llive/issues"
Family = "https://github.com/furuse-kazufumi/llmesh-suite"

[project.scripts]
llive = "llive.cli.main:app"
llive-demo = "llive.demo.runner:main"
llive-manifest = "llive.fullsense.manifest:main"

[project.entry-points."llive.subblocks"]
# builtin sub-blocks register themselves via llive.container.subblocks.builtin.register_builtins()
# external plugins can advertise their entry-points here

[tool.hatch.build.targets.wheel]
packages = ["src/llive"]
# src/llive/_specs/ is automatically included because it lives under the package tree

[tool.hatch.build.targets.sdist]
include = [
  "src/llive",
  "specs",
  "docs",
  "scripts",
  "tests",
  "crates",
  "CHANGELOG.md",
  "README.md",
  "LICENSE",
  "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = [
  "E501",   # line too long (handled by formatter)
  "B008",   # typer.Option / typer.Argument idiomatic in defaults
  "RUF002", # full-width em-dash / hyphens are intentional in Japanese docstrings
  "RUF003", # likewise for ambiguous Unicode in comments
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "gpu: requires GPU",
  "slow: slow test, skip in CI",
  "integration: end-to-end integration test",
]
addopts = "-ra --strict-markers"

[tool.coverage.run]
source = ["src/llive"]
omit = [
  "tests/*",
  "*/__init__.py",
  "src/llive/core/adapter.py",  # HFAdapter requires torch (entirely optional path)
]

[tool.coverage.report]
# 99% target: exclude defensive, environment-dependent, and optional-dep paths
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  # Defensive guards that only fire on misuse / corrupted state
  "raise FileNotFoundError",
  "raise KeyError",
  "raise RuntimeError",
  "raise ValueError",
  "raise ChangeOpError",
  "raise NestedContainerError",
  "raise SchemaValidationError",
  "raise WikiSchemaError",
  # Environment-detection fallbacks
  "return Path\\(",
  "if env:",
  # Idiomatic empty-tail guards
  "if not\\s+\\w+:\\s*$",
  "if not isinstance",
  # Protocol methods
  "\\.\\.\\.\\s*$",
  # Fallback returns
  "return self\\._fallback_dim",
  "return None\\s*$",
  "return\\s+idx\\s*$",
  "return\\s+default\\s*$",
  # Linker / Closer / context-manager exits with no logic
  "def close\\(self\\)",
  "def __exit__\\(",
  "def __enter__\\(",
  "if \"already exists\"",
  "for parent in here\\.parents",
  # `continue` short circuits
  "^\\s+continue\\s*$",
  # Default-arg shortcut helpers
  "def _default_",
  # Optional-dep ModuleNotFoundError handlers
  "except ModuleNotFoundError",
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
